Put some of the definitions back
