+++ link-grammar/sat-solver/word-tag.cpp
@@ -3,10 +3,11 @@
 
 extern "C" {
 #ifdef DEBUG
-#include <build-disjuncts.h>      // for prt_exp_mem()
-#include <dict-api.h>             // for print_expression()
+#include <dict-common/build-disjuncts.h>      // for prt_exp_mem()
+#include <dict-comm0n/dict-api.h>             // for print_expression()
 #endif
 #include "error.h"
+#include "tokenize/tok-structures.h"
 #include "utilities.h"
 }
