+++ link-grammar/tokenize/tok-structures.h
@@ -15,14 +15,6 @@
 #ifndef _TOK_STRUCTURES_H_
 #define _TOK_STRUCTURES_H_
 
-/* Word subscripts come after the subscript mark (ASCII ETX)
- * In the dictionary, a dot is used; but that dot interferes with dots
- * in the input stream, and so we convert dictionary dots into the
- * subscript mark, which we don't expect to see in user input.
- */
-#define SUBSCRIPT_MARK '\3'
-#define SUBSCRIPT_DOT '.'
-
 /* Suffixes start with it.
  * This is needed to distinguish suffixes that were stripped off from
  * ordinary words that just happen to be the same as the suffix.
@@ -68,28 +60,6 @@ struct gword_set
 	struct gword_set *chain_next;
 };
 
-/**
- * Word, as represented shortly after tokenization, but before parsing.
- *
- * X_node* x:
- *    Contains a pointer to a list of expressions from the dictionary,
- *    Computed by build_sentence_expressions().
- *
- * Disjunct* d:
- *   Contains a pointer to a list of disjuncts for this word.
- *   Computed by: prepare_to_parse(), but modified by pruning and power
- *   pruning.
- */
-struct Word_struct
-{
-	const char *unsplit_word;
-
-	X_node * x;          /* Sentence starts out with these, */
-	Disjunct * d;        /* eventually these get generated. */
-	bool optional;       /* Linkage is optional. */
-
-	const char **alternatives;
-};
 
 typedef enum
 {
